Rename register confirmation file
